fix creation and deletion timestamp for work metadata. (#159)
* fix creation and deletion timestamp for work metadata.


* address comments.


---------

+++ pkg/services/resource.go
@@ -4,11 +4,13 @@ import (
 	"context"
 	"fmt"
 	"reflect"
+	"time"
 
 	cloudeventstypes "github.com/cloudevents/sdk-go/v2/types"
 	"github.com/openshift-online/maestro/pkg/dao"
 	"github.com/openshift-online/maestro/pkg/db"
 	logger "github.com/openshift-online/maestro/pkg/logger"
+	"gorm.io/gorm"
 
 	cegeneric "open-cluster-management.io/sdk-go/pkg/cloudevents/generic"
 	cetypes "open-cluster-management.io/sdk-go/pkg/cloudevents/generic/types"
@@ -23,7 +25,7 @@ type ResourceService interface {
 	Create(ctx context.Context, resource *api.Resource) (*api.Resource, *errors.ServiceError)
 	Update(ctx context.Context, resource *api.Resource) (*api.Resource, *errors.ServiceError)
 	UpdateStatus(ctx context.Context, resource *api.Resource) (*api.Resource, bool, *errors.ServiceError)
-	MarkAsDeleting(ctx context.Context, id string) *errors.ServiceError
+	MarkAsDeleting(ctx context.Context, id string, deletionTimestamp time.Time) *errors.ServiceError
 	Delete(ctx context.Context, id string) *errors.ServiceError
 	All(ctx context.Context) (api.ResourceList, *errors.ServiceError)
 
@@ -214,16 +216,48 @@ func (s *sqlResourceService) UpdateStatus(ctx context.Context, resource *api.Res
 	return updated, true, nil
 }
 
-// MarkAsDeleting marks the resource as deleting by setting the delete_at timestamp.
+// MarkAsDeleting marks the resource as deleting by setting the deletion timestamp.
+// There are two cases:
+// 1. If deletionTimestamp is not provided, the resource will be deleted and the deleteAt
+// will be set by the database.
+// 2. If deletionTimestamp is provided, the resource will be marked as deleting in both
+// the deleteAt field and deletionTimestamp in the work metadata before deletion.
+//
 // The Resource Deletion Flow:
 // 1. User requests deletion
 // 2. Maestro marks resource as deleting by soft delete, adds delete event to DB
 // 3. Maestro handles delete event and sends CloudEvent to work-agent
 // 4. Work-agent deletes resource, sends CloudEvent back to Maestro
 // 5. Maestro hard deletes resource from DB
-func (s *sqlResourceService) MarkAsDeleting(ctx context.Context, id string) *errors.ServiceError {
-	if err := s.resourceDao.Delete(ctx, id, false); err != nil {
-		return handleDeleteError("Resource", errors.GeneralError("Unable to delete resource: %s", err))
+func (s *sqlResourceService) MarkAsDeleting(ctx context.Context, id string, deletionTimestamp time.Time) *errors.ServiceError {
+	if deletionTimestamp.IsZero() {
+		// If deletionTimestamp is not provided, delete the resource immediately.
+		// The database deletion mechanism ensures the deleteAt field of the resource is set.
+		if err := s.resourceDao.Delete(ctx, id, false); err != nil {
+			return handleDeleteError("Resource", errors.GeneralError("Unable to delete resource: %s", err))
+		}
+	} else {
+		// If deletionTimestamp is provided, the resource will be marked as deleting in both
+		// the deleteAt field and deletionTimestamp in the work metadata before deletion.
+		found, err := s.resourceDao.Get(ctx, id)
+		if err != nil {
+			return handleGetError("Resource", "id", id, err)
+		}
+		// set deletionTimestamp in work metadata
+		workMetaValue, ok := found.Payload["metadata"]
+		if ok {
+			workMeta, ok := workMetaValue.(map[string]interface{})
+			if ok {
+				workMeta["deletionTimestamp"] = deletionTimestamp.Format(time.RFC3339)
+				found.Payload["metadata"] = workMeta
+			}
+		}
+		// set deletedAt in resource meta
+		found.DeletedAt = gorm.DeletedAt{Time: deletionTimestamp, Valid: true}
+		_, err = s.resourceDao.Update(ctx, found)
+		if err != nil {
+			return handleUpdateError("Resource", err)
+		}
 	}
 
 	if _, err := s.events.Create(ctx, &api.Event{
